# bpo-38436: Improved performance for list addition.