--- a/Misc/NEWS.d/next/Core and Builtins/2019-10-10-19-20-24.bpo-38436.LfVbc-.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2019-10-10-19-20-24.bpo-38436.LfVbc-.rst
@@ -1,0 +1,1 @@
+Improved the performance of repeated/literal :class:`list` addition.