@pablogsal 2019-10-10
Where the benchmarks executed with CPU isolation? I am a bit suspicious of the 

```
- pickle_list: 3.63 us +- 0.08 us -> 3.78 us +- 0.10 us: 1.04x slower (+4%)
```

If not, can you repeat them with CPU isolation and affinity? Check [this](https://vstinner.github.io/perf-visualize-system-noise-with-cpu-isolation.html) for more info