@brandtbucher 2019-10-10
@pablogsal No isolation (I'm on a MacBook) but I can rerun with affinity on two cores.