@pablogsal 2019-10-10
> @pablogsal No isolation (I'm on a MacBook) but I can rerun with affinity on two cores.

I can try to run the benchmarks on my machine or the speed.python.org server.